api amifiles/requirements.txt
fastapi
uvicorn
azure-storage-blob